requirements.txt
-e git+https://github.com/yono/python-extractcontent#egg=extractcontent
dragnet
eatiht
goose-extractor
readability-lxml
